Rename ./client/webpack to avoid confusion with `node_modules/webpack`
