+++ client/webpack-utils/extensions.js
